Reimplemented the client
Before, the client was created as a testing rig for the server and was
programmed very rigidly, in such a way that would make it difficult to
extend functionality beyond using it as a test rig. Now the client has
a simple "source destination" command-line interface for choosing the
name of your source and destination files, although it still only
supports client -> server transfers.